Make {csr|csc}_from_dense methods of CsMatI
